+++ lib/Phpfastcache/Drivers/Arangodb/Event.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Phpfastcache\Drivers\Arangodb;
+
+class Event extends \Phpfastcache\Event\Event
+{
+    public const ARANGODB_CONNECTION = 'ArangodbConnection';
+    public const ARANGODB_COLLECTION_PARAMS = 'ArangodbCollectionParams';
+}