+++ lib/Phpfastcache/Drivers/Dynamodb/Event.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Phpfastcache\Drivers\Dynamodb;
+
+class Event extends \Phpfastcache\Event\Event
+{
+    public const DYNAMODB_CREATE_TABLE = 'DynamodbCreateTable';
+}