+++ lib/Phpfastcache/Event/EventManagerInterface.php
@@ -45,6 +45,7 @@ use Phpfastcache\Exceptions\PhpfastcacheInvalidArgumentException;
  * @method Void onArangodbConnection(Callable $callable, ?string $callbackName = null)
  * @method Void onArangodbCollectionParams(Callable $callable, ?string $callbackName = null)
  * @method Void onDynamodbCreateTable(Callable $callable, ?string $callbackName = null)
+ * @method Void onSolrBuildEndpoint(Callable $callable, ?string $callbackName = null)
  */
 interface EventManagerInterface
 {